Quiet maybe uninit warning
When doing a debug build, gcc spit out a maybe used unitialized warning.
No idea if that's right... but it's basically free to just initialize
it. Failure is probably safe one (and if its ever returned, there was a
bug in the original, it would have previously been undefined behavior).