Recover from loss of ip request reply
Previously, if the reply packet to an IP request (MODECFG_R0/R1) were
lost, the modecfg client (requestor) would retransmit it, but then the
modecfg server would discard it, because it wasn't the expected packet.
The connection thus hung until it eventually timed out and terminated.

Now, instead, retransmit the reply upon receiving a duplicate request.
This allows the connection to succeed.

Also up the maximum number of times we'll retransmit in response to a
duplicate packet from 2 to 4. That allows for a longer blip and the
connection to still succeed.

Fixes internal issue [TABLET-1418]