Delete the modecfg retransmit on phase2 start
Previously, even once we got the response from the tablet, we kept
retransmitting the modecfg response... which meant that within a minute
or two we'd run out of retransmits and terminate the connection.

This is due how the timers work at the phase1 → phase2 transition. Work
around it well enough for now.

Find a better fix once we upgrade to a newer Libreswan.

Internal ticket [TABLET-1418].